@@ -8,7 +8,7 @@
         }
     ],
     "require": {
-        "php": ">=8.0"
+        "php": "^8.1"
     },
     "require-dev": {
         "phpstan/phpstan": "^1.2.0",
